Created BaseDAO, ProductDAO and more
Started separating the data access tier from the services. Created a
BaseDAO (data access object) that will be used as a parent to all DAOs
and provides all common functionality for the database interaction.
ProductDAO implements more specific behaviour for the products data
operations. Modified the ShopService to use the ProductDAO.